feat: Add remove-game command to remove a registered game
This commit adds the remove-game command to the configuration module. The command allows users to remove a registered game from the list of watched games. It includes the necessary options for specifying the name of the game to be removed. This addition enhances the functionality of the bot by providing users with the ability to manage their watched games more effectively.