feat: Add get-game command to retrieve game information
This commit adds the get-game command to the configuration module. The command allows users to retrieve information about a registered game. It includes the necessary options for specifying the name of the game to get information about. This addition enhances the functionality of the bot by providing users with the ability to easily access game details.