feat: Add here command to set news channel
The code changes add a new command called "here" to the configuration module. This command allows users to set the news channel where the bot will spread news. It retrieves the channel ID from the interaction and updates the configuration accordingly. This addition enhances the functionality of the bot by providing a convenient way for users to specify the news channel.